python/jude/v3/requirements.txt
gunicorn==20.1.0
uvicorn[standard]==0.21.1
psycopg==3.1.8
psycopg-c==3.1.8
psycopg-pool==3.1.7
python-rapidjson==1.10